refactor: move group switcher files into a GroupSwitcher folder
ActiveGroupSwitcher.tsx, ScopeMenuBody.tsx, and scopeDisplay.ts are
used exclusively by the group switcher, so group them together
instead of leaving them loose alongside unrelated AppHeader files.
