+++ src/components/layout/AppHeader/GroupSwitcher/ActiveGroupSwitcher.tsx
