+++ src/components/layout/AppHeader/GroupSwitcher/scopeDisplay.ts
